+++ client/src/components/billing/checkout/checkoutForm.js
@@ -1,24 +1,24 @@
-import React from 'react';
-import { injectStripe } from 'react-stripe-elements';
-import CardSection from './cardSection';
+import React from 'react'
+import { injectStripe } from 'react-stripe-elements'
+import CardSection from './CardSection'
 
 class CheckoutForm extends React.Component {
-    handleSubmit = (ev) => {
-        ev.preventDefault();
+  handleSubmit = ev => {
+    ev.preventDefault()
 
-        this.props.stripe.createToken({ name: 'Random Name'}). then(({token})=> {
-            console.log('Received Stripe Token:' , token);
-        });
-    }
+    this.props.stripe.createToken({ name: 'Random Name' }).then(({ token }) => {
+      console.log('Received Stripe Token:', token)
+    })
+  }
 
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-            <CardSection />
-            <button>PAY</button>
-            </form>
-        );
-    }
+  render() {
+    return (
+      <form onSubmit={this.handleSubmit}>
+        <CardSection />
+        <button>PAY</button>
+      </form>
+    )
+  }
 }
 
-export default injectStripe(CheckoutForm);
+export default injectStripe(CheckoutForm)
