+++ client/src/components/billing/checkout/stripeCheckout.js
@@ -1,15 +1,15 @@
-import React from 'react';
-import { Elements } from 'react-stripe-elements';
-import CheckoutForm from './checkoutForm';
+import React from 'react'
+import { Elements } from 'react-stripe-elements'
+import CheckoutForm from './CheckoutForm'
 
 class Checkout extends React.Component {
-    render() {
-        return (
-            <Elements>
-                <CheckoutForm />
-            </Elements>
-        );
-    }
+  render() {
+    return (
+      <Elements>
+        <CheckoutForm />
+      </Elements>
+    )
+  }
 }
 
-export default Checkout;
+export default Checkout
